Fix some false positives with mind control.
