+++ src/game/Anticheat/MovementAnticheat/MovementAnticheat.cpp
@@ -874,7 +874,7 @@ bool MovementAnticheat::CheckNoFallTime(MovementInfo const& movementInfo, uint16
         return false;
     }
 
-    if (!GetLastMovementInfo().ctime ||
+    if (!GetLastMovementInfo().WasSentBySession(m_session->GetGUID()) ||
         !GetLastMovementInfo().HasMovementFlag(MOVEFLAG_JUMPING | MOVEFLAG_FALLINGFAR))
     {
         m_jumpFlagCount = 0;
@@ -916,7 +916,7 @@ bool MovementAnticheat::CheckFallReset(MovementInfo const& movementInfo) const
     if (movementInfo.HasMovementFlag(MOVEFLAG_FIXED_Z))
         return false;
 
-    if (GetLastMovementInfo().ctime)
+    if (GetLastMovementInfo().WasSentBySession(m_session->GetGUID()))
     {
         if (!GetLastMovementInfo().HasMovementFlag(MOVEFLAG_JUMPING | MOVEFLAG_FALLINGFAR))
             return true;
@@ -930,7 +930,7 @@ bool MovementAnticheat::CheckFallStop(MovementInfo const& movementInfo, uint16 o
     if (!sWorld.getConfig(CONFIG_BOOL_AC_MOVEMENT_CHEAT_BAD_FALL_STOP_ENABLED))
         return false;
 
-    if (!GetLastMovementInfo().ctime)
+    if (!GetLastMovementInfo().WasSentBySession(m_session->GetGUID()))
         return false;
 
     if (IsFallEndOpcode(opcode))
@@ -957,7 +957,7 @@ bool MovementAnticheat::CheckMoveStart(MovementInfo const& movementInfo, uint16 
     if (!sWorld.getConfig(CONFIG_BOOL_AC_MOVEMENT_CHEAT_BAD_MOVE_START_ENABLED))
         return false;
 
-    if (!GetLastMovementInfo().ctime)
+    if (!GetLastMovementInfo().WasSentBySession(m_session->GetGUID()))
         return false;
 
     if (IsFallEndOpcode(opcode))
@@ -1082,7 +1082,7 @@ uint32 MovementAnticheat::CheckTimeDesync(MovementInfo const& movementInfo)
     uint32 cheatFlags = 0x0;
 #define APPEND_CHEAT(t) cheatFlags |= (1 << t)
 
-    if (GetLastMovementInfo().ctime)
+    if (GetLastMovementInfo().WasSentBySession(m_session->GetGUID()))
     {
         if (GetLastMovementInfo().moveFlags & MOVEFLAG_MASK_MOVING)
         {
